bot/pyproject.toml
[project]
name = "se-toolkit-bot"
version = "0.1.0"
description = "SE Toolkit Lab 7 - Telegram Bot for LMS"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["*.py", "handlers/", "services/"]
